feat: Fix FSDP model saving code

+++ src/tools/training_logger.py
@@ -101,36 +101,39 @@ class TrainingLogger:
             LOGGER.info(f"{colorstr('green', 'bold', ', '.join(msg))}\n")
 
     
-    def delete_file(self, save_dir, flag):
-        file = list(filter(lambda x: flag in x, os.listdir(save_dir)))
-        if len(file) > 0:
-            os.remove(os.path.join(save_dir, file[0]))
+    def delete_file(self, save_dir, flag, is_rank_zero):
+        if is_rank_zero:
+            file = list(filter(lambda x: flag in x, os.listdir(save_dir)))
+            if len(file) > 0:
+                os.remove(os.path.join(save_dir, file[0]))
 
 
-    def save_model(self, save_dir, model):
-        if not hasattr(self, 'validation_epoch_result') or len(self.validation_epoch_result) == 0:
-            LOGGER.warning(f'{colorstr("red", "No log data to save")}')
-            return
+    def save_model(self, save_dir, model, is_rank_zero, is_fsdp=False):
+        if is_rank_zero or is_fsdp:
+            if not hasattr(self, 'validation_epoch_result') or len(self.validation_epoch_result) == 0:
+                LOGGER.warning(f'{colorstr("red", "No log data to save")}')
+                return
 
-        epoch, step = self.log_data['epoch'][-1], self.log_data['step'][-1]
-        lower_flag, higher_flag = self.model_manager.update_best(self.validation_epoch_result)
+            epoch, step = self.log_data['epoch'][-1], self.log_data['step'][-1]
+            lower_flag, higher_flag = self.model_manager.update_best(self.validation_epoch_result)
 
-        if lower_flag:
-            self.delete_file(save_dir, 'loss')
-            model_path = os.path.join(save_dir, f'model_epoch:{epoch}_step:{step}_loss_best.pt')
-            self.model_manager.save(model, model_path, self.validation_epoch_result)
+            if lower_flag:
+                self.delete_file(save_dir, 'loss', is_rank_zero)
+                model_path = os.path.join(save_dir, f'model_epoch:{epoch}_step:{step}_loss_best.pt')
+                self.model_manager.save(model, model_path, self.validation_epoch_result, is_rank_zero)
 
-        if higher_flag:
-            self.delete_file(save_dir, 'metric')
-            model_path = os.path.join(save_dir, f'model_epoch:{epoch}_step:{step}_metric_best.pt')
-            self.model_manager.save(model, model_path, self.validation_epoch_result)
-        
-        self.delete_file(save_dir, 'last')
-        model_path = os.path.join(save_dir, f'model_epoch:{epoch}_step:{step}_last_best.pt')
-        self.model_manager.save(model, model_path, self.validation_epoch_result)
+            if higher_flag:
+                self.delete_file(save_dir, 'metric', is_rank_zero)
+                model_path = os.path.join(save_dir, f'model_epoch:{epoch}_step:{step}_metric_best.pt')
+                self.model_manager.save(model, model_path, self.validation_epoch_result, is_rank_zero)
+            
+            self.delete_file(save_dir, 'last', is_rank_zero)
+            model_path = os.path.join(save_dir, f'model_epoch:{epoch}_step:{step}_last_best.pt')
+            self.model_manager.save(model, model_path, self.validation_epoch_result, is_rank_zero)
 
     
-    def save_logs(self, save_dir):
-        self.delete_file(save_dir, 'log_data')
-        with open(os.path.join(save_dir, 'log_data.pkl'), 'wb') as f:
-            pickle.dump(self.log_data, f)
+    def save_logs(self, save_dir, is_rank_zero):
+            self.delete_file(save_dir, 'log_data', is_rank_zero)
+            if is_rank_zero:
+                with open(os.path.join(save_dir, 'log_data.pkl'), 'wb') as f:
+                    pickle.dump(self.log_data, f)
